Extract goal tag helper functions
